refactor: :recycle: Refactor gatsby-cli installation
Functions have been moved to an installer in lib and proper handlers have been defined in the api. The slice has been removed as it does not serve much of a purpose.

The handlers in the api were defined as we will be checking the installation and firing it from the UI instead of the main thread as we will want to show a loading screen when that happens. Something similair to the auth progress one.