fix: :bug: Do not crash the app when default path is invalid
The FirstTime view will still allow to set a bad path but will show a notification that it is invalid. The local publications will simply not be loaded if the path is invalid.

Validation of the Firsttime view input will also be added so that the button in the view can be disabled.