fix: :pushpin: restrict ignore to a specific version
for some reason isomorphic-git throws an error (ign.test is not a function) when using the statusMatrix method. The .test method was introduced in the 5.0.0 version of ignore package. Pinning the dependency for all other packages is a rough fix, as it is unclear why this happens - theoretically isomorphic-git has a correct dependecy and references it correctly, but it gets messed up at runtime.