fix: :passport_control: Authorize user in the packaged app
Authorization callback url has been moved to a new default protocol for the app (publab://auth/callback). Environmental variables are now also properly loaded at webpack build so that they are available in the packaged app.

BREAKING: The callback url in GitHub has to be set to publab://auth/callback for auth to work